requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
dj-database-url>=2.2
psycopg[binary]>=3.1
markdown>=3.6
django-markdownx>=4.0
whitenoise>=6.6
openai>=1.40
requests>=2.32
django-redis>=5.4
django-csp>=3.8
celery>=5.3
redis>=5.0
